+++ android/app/src/main/java/com/yaskur/belajarmembaca/MainActivity.java
@@ -1,0 +1,5 @@
+package com.yaskur.belajarmembaca;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}